# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Simple CRUD User tanpa DB (In-Memory)",
        "title": "User API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/examples/400": {
            "get": {
                "description": "Demonstrates when client sends invalid data",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "error-examples"
                ],
                "summary": "Example of 400 Bad Request error",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Send 'invalid' to trigger error",
                        "name": "invalid",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/examples/401": {
            "get": {
                "description": "Demonstrates when authentication is required but not provided",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "error-examples"
                ],
                "summary": "Example of 401 Unauthorized error",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/examples/403": {
            "get": {
                "description": "Demonstrates when user is authenticated but doesn't have permission",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "error-examples"
                ],
                "summary": "Example of 403 Forbidden error",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User role (admin or user)",
                        "name": "role",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/examples/404": {
            "get": {
                "description": "Demonstrates when requested resource doesn't exist",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "error-examples"
                ],
                "summary": "Example of 404 Not Found error",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Resource ID",
                        "name": "id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/examples/409": {
            "get": {
                "description": "Demonstrates when there's a conflict with current state",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "error-examples"
                ],
                "summary": "Example of 409 Conflict error",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Email to check",
                        "name": "email",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/examples/422": {
            "get": {
                "description": "Demonstrates when server understands the request but can't process it",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "error-examples"
                ],
                "summary": "Example of 422 Unprocessable Entity error",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User age",
                        "name": "age",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/examples/429": {
            "get": {
                "description": "Demonstrates rate limiting",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "error-examples"
                ],
                "summary": "Example of 429 Too Many Requests error",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Number of requests made",
                        "name": "requests",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/examples/500": {
            "get": {
                "description": "Demonstrates unexpected server error",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "error-examples"
                ],
                "summary": "Example of 500 Internal Server Error",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Set to 'error' to trigger 500",
                        "name": "trigger",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/examples/503": {
            "get": {
                "description": "Demonstrates when service is temporarily unavailable",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "error-examples"
                ],
                "summary": "Example of 503 Service Unavailable error",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Set to 'true' for maintenance mode",
                        "name": "maintenance",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
    }
}